mito-ai: refactor to match codebase
